break up page into components
